+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 <!--next-version-placeholder-->
+Bug fix: Executable content files where returning bytes instead of strings, this is python3 behaviour and has been fixed
 
 ## v3.1.1 - 2021-12-14