+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 <!--next-version-placeholder-->
 Bug fix: Executable content files where returning bytes instead of strings, this is python3 behaviour and has been fixed
+Bug fix: Page header preserve file name now is enforced, before the file name could have the default extension added, if specified in site.conf'
+
 
 ## v3.1.1 - 2021-12-14