+++ CHANGELOG.md
@@ -1,10 +1,12 @@
 # Changelog
 
 <!--next-version-placeholder-->
-Bug fix: Executable content files where returning bytes instead of strings, this is python3 behaviour and has been fixed
-Bug fix: Page header preserve file name now is enforced, before the file name could have the default extension added, if specified in site.conf'
+Fixed: Executable content files where returning bytes instead of strings, this is python3 behaviour and has been fixed
+Fixed: Page header preserve file name now is enforced, before the file name could have the default extension added, if specified in site.conf'
 Feature: Added page header Exclude html minify, set to True to disable html minifying of the page
-Bug fix: Multiple variable in a line containing : where matching too much
+Fixed: Multiple variable in a line containing : where matching too much
+Changed: Moved deploy_scripts to new deploy hook
+
 
 ## v3.1.1 - 2021-12-14